separate Example to Part1 - 3
