+++ Examples/Scripts/PrefsGUIExample_Part1.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 61933781acb3fb14faf7443efab03b2c
+timeCreated: 1473738440
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
