+++ Examples/Scripts/PrefsGUIExample_Part2.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 84e9b87159a01f64da8cb763532f1382
+timeCreated: 1473738440
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
