+++ Examples/Scripts/PrefsGUIExample_Part3.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 99088179b88fc1d46adc2fc60ac2fbee
+timeCreated: 1473738440
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
