+++ Runtime/MaterialPropertyAccessor/Sample/Sample.mat
@@ -72,12 +72,12 @@ Material:
     - _Mode: 0
     - _OcclusionStrength: 1
     - _Parallax: 0.02
-    - _SmoothnessTextureChannel: 0
+    - _SmoothnessTextureChannel: 0.32978722
     - _SpecularHighlights: 1
     - _SrcBlend: 1
     - _UVSec: 0
     - _ZWrite: 1
     m_Colors:
     - <noninit>: {r: 0, g: 1, b: 1, a: 1}
-    - _Color: {r: 0, g: 1, b: 1, a: 1}
-    - _EmissionColor: {r: 0, g: 0, b: 0.5, a: 1}
+    - _Color: {r: 0.21306995, g: 0.4875, b: 0.4875, a: 1}
+    - _EmissionColor: {r: 0.10052297, g: 0.10052297, b: 0.38035715, a: 1}
